changed type on field v on TeethDistribution

+++ db/migrate/20201124202740_create_teeth_distributions.rb
@@ -2,7 +2,7 @@ class CreateTeethDistributions < ActiveRecord::Migration[6.0]
   def change
     create_table :teeth_distributions do |t|
       t.belongs_to :smoothness_norm, foreign_key: true, type: :integer
-      t.integer :v, null: false
+      t.float :v, null: false
       t.float :value, null: false
 
       t.timestamps
